Cache successful authentication for 1 hour.
Create a cache file in user's home directory which should hold the source IP from which the user connected and the current timestamp after successful authentication and 2FA code verification. On next authentication check the cache file and skip verification code for the same user and source IP address if the cached entry has not expired.